Add files via upload
This is version V2.2 of the Exchange Database statistics report. Based on thresholds in settings.cfg, report includes now a critical CSV and HTML table and staging CSV (List of databases, whoch can be used as staging databases for mailbox moves). Read the description within the script carefully. NOTICE: If you were using older versions of this script, please dowload and/or update your SETTINGS.CFG, as there are new and renamed values in it.